Add search bar and init.php
Added search bar and some styling for it to look suave and sexy.
It will be hooked up to a Fuse IV, to get fluids and fuzzy search
results.

init.php is likely to be removed, but I wanted to add it anyways
so you all can see how bad I am at writing PHP.